`:to :llvm` emits LLVM IR and needs no linker to do it -- `clang` links, against
a new `nativeabi/` staticlib. `:to :native` still refuses, now for its own
reason rather than sharing one that was only ever true of it.

TWO SEMANTIC MERGES, neither mechanical:

`src/flint/selfhost.cljc` -- the branch extracted `build-image`, replacing two
copies of the resolve-then-compile path with one. Taking it wholesale would
have dropped `:dialect` and `:prelude` from the source map, because the branch
predates them and that map is rebuilt field by field. Their refactor plus those
fields; verified afterwards that a prelude name still resolves with no
`:require`, which is precisely what a silent drop would have broken.

`DECISIONS.md` -- three agents appended sections at EOF. Both sides kept.

AND A GATE THAT CANNOT PASS ON A FRESH CLONE, found by the agent and confirmed
here by deleting `sdks/esm/dist` and watching `test/capability.clj` fail. It
compiles against `sdks/esm/dist/flint.js`, which is generated and gitignored,
and `bin/test` builds that 179 lines LATER. Every tree that has ever built the
SDK hides it -- which is every developer's and none of CI's.
`.github/workflows/test.yml` runs `./bin/test` on a fresh checkout with no SDK
step, on pull requests only, so nothing has opened one since the dependency
landed and nothing has reported it.

Built on demand rather than by reordering, because the SDK section tests the
DISTRIBUTABLE and has to keep running where it does. Verified both ways: the
suite fails without `dist`, and passes with the bootstrap.

+++ runtime/src/mem.rs
@@ -332,9 +332,16 @@ impl Space {
     }
     /// Where object addresses are measured from. Compiled code needs it to read
     /// a slot without a call back into Rust (`DECISIONS.md#emit-wasm-instead-of-dispatch`).
+    ///
+    /// `usize`, not `u32`. This is a HOST pointer -- `ptr()` two functions up
+    /// adds an `Addr` to it -- and the `as u32` that used to be here was the
+    /// same truncation `Value::as_heap` already has a comment about: right
+    /// while the only reader was a wasm module, and a wild base everywhere
+    /// else. The LLVM target is the everywhere else, and the symptom was a
+    /// `SIGSEGV` on the first `:upval` (`DECISIONS.md#llvm-ir-target`).
     #[inline(always)]
-    pub fn base_addr(&self) -> u32 {
-        self.base as u32
+    pub fn base_addr(&self) -> usize {
+        self.base as usize
     }
     #[inline(always)]
     pub fn read_u64(&self, addr: Addr) -> u64 {
